backend/requirements.txt
flask
flask-cors
psycopg2-binary
flask_sqlalchemy
python-dotenv
cloudinary